Move both files in Sofa.Type
